+++ src/main/java/sunsetsatellite/catalyst/core/interfaces/mixins/IKeybinds.java
@@ -1,0 +1,8 @@
+package sunsetsatellite.catalyst.core.interfaces.mixins;
+
+import net.minecraft.client.option.BooleanOption;
+
+public interface IKeybinds {
+
+	BooleanOption getNetworkRenderOption();
+}