+++ src/main/java/sunsetsatellite/catalyst/core/util/mixin/interfaces/UnlimitedItemStack.java
@@ -1,0 +1,9 @@
+package sunsetsatellite.catalyst.core.util.mixin.interfaces;
+
+public interface UnlimitedItemStack {
+    void setUnlimited(boolean unlimited);
+
+	void enableCustomMaxSize(int maxSize);
+
+	void disableCustomMaxSize();
+}