+++ all/CHANGELOG.md
@@ -1,1 +1,6 @@
-- Various bugfixes
+- Added support for fluids and multipart in Too Many Blocks.
+- Removed the old energy API.
+- All modules now have their own options category. 
+- Added a way to control ItemStack sizes. 
+- Block network fixes.
+  Added a way of rendering block networks for debugging purposes.